+++ src/dataset.rs
@@ -552,7 +552,7 @@ impl Dataset {
     ///
     /// ```
     /// # use gdal::Driver;
-    /// # let driver = Driver::get("GPKG").unwrap();
+    /// # let driver = Driver::get_by_name("GPKG").unwrap();
     /// # let mut dataset = driver.create_vector_only("/vsimem/example.gpkg").unwrap();
     /// let blank_layer = dataset.create_layer(Default::default()).unwrap();
     /// ```
@@ -562,7 +562,7 @@ impl Dataset {
     /// ```
     /// # use gdal::{Driver, LayerOptions};
     /// # use gdal::spatial_ref::SpatialRef;
-    /// # let driver = Driver::get("GPKG").unwrap();
+    /// # let driver = Driver::get_by_name("GPKG").unwrap();
     /// # let mut dataset = driver.create_vector_only("/vsimem/example.gpkg").unwrap();
     /// let roads = dataset.create_layer(LayerOptions {
     ///     name: "roads",
@@ -720,7 +720,7 @@ impl Dataset {
     /// }
     /// #
     /// # fn main() -> gdal::errors::Result<()> {
-    /// #     let driver = gdal::Driver::get("SQLite")?;
+    /// #     let driver = gdal::Driver::get_by_name("SQLite")?;
     /// #     let mut dataset = driver.create_vector_only(":memory:")?;
     /// #     create_point_grid(&mut dataset)?;
     /// #     assert_eq!(dataset.layer(0)?.features().count(), 10000);